Removed CharacterDefinition
Changed Container slot type from u8 to u16
Added basic tracking of container slots
Added new character prepared statements (CharacterGetInfoByID, CharacterGetInfoByName, CharacterGetItemInstances, CharacterGetItemsInContainer)
Added new character persistence util (CharacterGetInfoByID, CharacterGetInfoByName, CharacterSetRaceGenderClass,
CharacterSetLevel, CharacterGetItems, CharacterGetItemsInContainer)
Removed item persistence util (LoadItemInstances)
Added basic Lua implementation
Added CharacterInfo
Added helper static constexpr variables for PlayerContainers
Added new tags (CharacterNeedsDeinitialization, CharacterNeedsContainerUpdate, CharacterNeedsDisplayUpdate)
Added MapID to Transform Components
Removed GameCommands (Old Persistence system)
Added CharacterUpdate System
Renamed DatabaseState -> GameCache
Added InitPlayer function to WorldState
Overhauled our persistence system
Added CacheUtil, ContainerUtil, NetworkUtil, CharacterUtil, ItemUtil, UnitUtil
Updated Submodule Engine

+++ Source/Server-Game/Server-Game/ECS/Util/MessageBuilderUtil.h
@@ -66,10 +66,10 @@ namespace ECS
 
         namespace Container
         {
-            bool BuildContainerCreate(std::shared_ptr<Bytebuffer>& buffer, u8 containerIndex, u32 itemID, GameDefine::ObjectGuid guid, const Database::Container& container);
-            bool BuildAddToSlot(std::shared_ptr<Bytebuffer>& buffer, u8 containerIndex, u8 slot, GameDefine::ObjectGuid itemGuid);
-            bool BuildRemoveFromSlot(std::shared_ptr<Bytebuffer>& buffer, u8 containerIndex, u8 slot);
-            bool BuildSwapSlots(std::shared_ptr<Bytebuffer>& buffer, u8 srcContainerIndex, u8 destContainerIndex, u8 srcSlot, u8 destSlot);
+            bool BuildContainerCreate(std::shared_ptr<Bytebuffer>& buffer, u16 containerIndex, u32 itemID, GameDefine::ObjectGuid guid, const Database::Container& container);
+            bool BuildAddToSlot(std::shared_ptr<Bytebuffer>& buffer, u16 containerIndex, u16 slot, GameDefine::ObjectGuid itemGuid);
+            bool BuildRemoveFromSlot(std::shared_ptr<Bytebuffer>& buffer, u16 containerIndex, u16 slot);
+            bool BuildSwapSlots(std::shared_ptr<Bytebuffer>& buffer, u16 srcContainerIndex, u16 destContainerIndex, u16 srcSlot, u16 destSlot);
         }
 
         namespace Spell
